requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numba==0.62.0
numpy==2.2.6
packaging==25.0
pandas==2.2.2
pillow==11.3.0
pynndescent==0.5.13
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.15.3
sentence-transformers==3.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.6
urllib3==2.5.0
